requirements.txt
numpy
tensorflow
streamlit
pillow